fixed(trusted0by):rename branch and component from about us to trusted-by
