+++ src/app/features/pages/trasted-by-us/trasted-by.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AboutUsComponent } from './about-us.component';
+import { AboutUsComponent } from './trasted-by.component';
 
 describe('AboutUsComponent', () => {
   let component: AboutUsComponent;
