+++ src/app/features/pages/trasted-by-us/trasted-by.component.ts
@@ -3,8 +3,8 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-about-us',
   imports: [],
-  templateUrl: './about-us.component.html',
-  styleUrl: './about-us.component.scss'
+  templateUrl: './trasted-by.component.html',
+  styleUrl: './trasted-by.component.scss'
 })
 export class AboutUsComponent {
   public images:string[]=['coconut.png','ginyard.png','habus.png','ingoude.png','ingoude2.png','velvet.png']
